// apps/web2/src/api/mods.ts
export const projectSchema = z.object({
  id: z.string(),
  slug: z.string().optional(),
  name: z.string(),
  summary: z.string().optional(),
  iconUrl: z.string().optional(),
  downloads: z.number().optional(),
});
// ...
export const searchMirrorSchema = z.object({
  provider: z.string(),
  projectId: z.string(),
  slug: z.string(),
  downloads: z.number(),
});
export const searchAllItemSchema = projectSchema.extend({
  provider: z.string(),
  mirror: searchMirrorSchema.optional(),
});
export type SearchAllItem = z.infer<typeof searchAllItemSchema>;